+++ src/components/Guilds/Filter/index.tsx
@@ -1,44 +1,18 @@
 import { useState } from 'react';
 import styled from 'styled-components';
 import { isDesktop, isMobile } from 'react-device-detect';
+import { useFilter } from 'contexts/Guilds/filters';
 
-import {
-  DropdownButton,
-  DropdownContent,
-  DropdownMenu,
-  DropdownPosition,
-} from '../common/DropdownMenu';
-
-import { FiChevronDown, FiCheck, FiArrowLeft } from 'react-icons/fi';
-
-import { Menu, MenuItem } from '../common/Menu';
-import { Button } from '../common/Button';
 import { InputText } from '../common/Form';
 import { Box } from '../common/Layout/Box';
 
-import { useMenu } from '../../../hooks/Guilds/useMenu';
-
-const DropdownMenuItem = styled(MenuItem)`
-  display: flex;
-  flex: 1;
-  justify-content: space-between;
-  cursor: pointer;
-
-  &:hover {
-    background-color: #f5f5f5;
-  }
-`;
+import { FilterMenu, FilterButton, FilterBadge } from './FilterMenu';
 
 const FilterContainer = styled(Box)`
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   margin-bottom: 1rem;
-`;
-
-const FilterButtons = styled.div`
-  display: flex;
-  flex-direction: row;
 `;
 
 const FilterRow = styled.div`
@@ -49,83 +23,25 @@ const FilterRow = styled.div`
   }
 `;
 
-const FilterButtonsComponent = () => {
-  const { isStatusSelected, isSchemeSelected, onToggleStatus, onToggleScheme } =
-    useMenu({ initialSchemas: [], initialStatuses: [] });
-
-  const [showScheme, setShowScheme] = useState(false);
-  const [showStatus, setShowStatus] = useState(false);
-
-  return (
-    <FilterButtons>
-      <DropdownMenu position={DropdownPosition.BottomRight}>
-        <DropdownButton
-          iconRight
-          onClick={() => {
-            setShowScheme(!showScheme);
-          }}
-        >
-          Scheme <FiChevronDown />
-        </DropdownButton>
-        <DropdownContent show={showScheme}>
-          <Menu>
-            {isMobile && <FiArrowLeft onClick={() => setShowScheme(false)} />}
-            <DropdownMenuItem onClick={() => onToggleScheme('a')}>
-              Schema 1 {isSchemeSelected('a') && <FiCheck />}
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => onToggleScheme('b')}>
-              Schema 2 {isSchemeSelected('b') && <FiCheck />}
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => onToggleScheme('c')}>
-              Schema 3 {isSchemeSelected('c') && <FiCheck />}
-            </DropdownMenuItem>
-          </Menu>
-        </DropdownContent>
-      </DropdownMenu>
-      <DropdownMenu>
-        <DropdownButton iconRight onClick={() => setShowStatus(!showStatus)}>
-          Status <FiChevronDown />
-        </DropdownButton>
-        <DropdownContent show={showStatus}>
-          <Menu>
-            {isMobile && <FiArrowLeft onClick={() => setShowStatus(false)} />}
-            <DropdownMenuItem onClick={() => onToggleStatus('a')}>
-              Status a {isStatusSelected('a') && <FiCheck />}
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => onToggleStatus('b')}>
-              Status b {isStatusSelected('b') && <FiCheck />}
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => onToggleStatus('c')}>
-              Status c {isStatusSelected('c') && <FiCheck />}
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => onToggleStatus('d')}>
-              Status d {isStatusSelected('d') && <FiCheck />}
-            </DropdownMenuItem>
-          </Menu>
-        </DropdownContent>
-      </DropdownMenu>
-    </FilterButtons>
-  );
-};
-
 export const Filter = () => {
   const [viewFilter, setViewFilter] = useState(false);
-
+  const { totalFilters } = useFilter();
   return (
     <FilterContainer>
       <FilterRow>
         <InputText placeholder="Proposal title" />
-        {isDesktop && <FilterButtonsComponent />}
+        {isDesktop && <FilterMenu />}
         {isMobile && (
-          <Button
+          <FilterButton
             onClick={() => setViewFilter(!viewFilter)}
-            active={viewFilter}
+            active={viewFilter || totalFilters > 0}
           >
             Filter
-          </Button>
+            {totalFilters > 0 && <FilterBadge>{totalFilters}</FilterBadge>}
+          </FilterButton>
         )}
       </FilterRow>
-      {isMobile && viewFilter && <FilterButtonsComponent />}
+      {isMobile && viewFilter && <FilterMenu />}
     </FilterContainer>
   );
 };
