+++ src/components/Guilds/common/Menu/index.tsx
@@ -11,4 +11,12 @@ export const MenuItem = styled.a`
   display: block;
   text-decoration: none;
   padding: 0.5rem 1rem;
+
+  &:visited,
+  &:active,
+  &:link {
+    text-decoration: none;
+    decoration: none;
+    color: inherit;
+  }
 `;
